move some vulnerability types from repocop to common for future use in obligatron
